fix: tool handlers
