+++ extension/src/lib/llm/tools/handlers/search.ts
@@ -1,8 +1,8 @@
-import { webSearch } from "./web-search";
+import { webSearch } from "./common";
 import type { SearchToolResult } from "../search";
 
 export const availableFunctions: {
   [key: string]: (args: any) => Promise<SearchToolResult>;
 } = {
-  webSearch: webSearch,
+  webSearch,
 };
